refactor(collections api): reutilización de lógica
